pyproject.toml
[project]
name = "learning-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "agno>=2.2.3",
    "groq>=0.33.0",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
    "tavily>=1.0.0",
    "tavily-python>=0.7.12",
    "yfinance>=0.2.66",
]
